convert BitFile.Size() to use uint64 rather than int64

+++ go/bitfile/BitFile.Size.go
@@ -13,13 +13,18 @@ import (
 )
 
 // Size - Return file size (in bytes)
-func (o *BitFile) Size() (sz int64, err error) {
+func (o *BitFile) Size() (sz uint64, err error) {
+
 	if o.file == nil {
 		return 0, fmt.Errorf("file unknown")
 	}
+
 	fileInfo, err := os.Stat(o.file.Name())
+
 	if err != nil {
 		return 0, err
 	}
-	return fileInfo.Size(), nil
+
+	return uint64(fileInfo.Size()), nil
+
 }
